// frontend/src/app/models/acquisition/supplier.ts
export interface Supplier {
  id:              number;
  name?:           string;
  address?:        string;
  mobile?:         string;
  fax?:            string;
  account?:        string;
  bank?:           string;
  mst?:            string;
  email?:          string;
  website?:        string;
  position?:       string;
  representative?: string;
  publicId?:       string;
  tenantName?:     string;
}